pyproject.toml
[project]
name = "once-upon-agentic-ai"
version = "1.0.0"
description = "A comprehensive workshop for mastering AI agent development with Strands framework"
requires-python = ">=3.10"
dependencies = [
    # Strands AI Framework
    "strands-agents>=1.0.0",
    "strands-agents-tools>=0.2.0",
    "strands-agents[a2a]",
    # Web Framework for APIs
    "fastapi[standard]",
    "pydantic",
    "uvicorn",
    # Configuration Management
    "python-dotenv",
    # Knowledge Base & Document Processing
    "chromadb>=0.4.0",
    "PyPDF2>=3.0.0",
    # Database
    "tinydb",
    # LangChain Integration
    "langchain_aws",
    "langchain_chroma",
    # Bedrock AgentCore
    "bedrock-agentcore",
    "bedrock-agentcore-starter-toolkit>=0.2.8",
]
